Uppgift 1: Fix bug in isValidInput() that prints min / max as floats.

+++ 02-Programmering-C-och-Cpp/IoT-25-CPP-Uppgift1/src/functions.cpp
@@ -183,7 +183,7 @@ bool isValidInput(const std::string& input, int typeSelector, float min, float m
             try {
                 int inputVal = std::stoi(input);
                 if (inputVal >= min && inputVal <= max) return true;
-                std::cout << "Please enter a number between " << min << " and " << max << "\n";
+                std::cout << "Please enter a number between " << static_cast<int>(min) << " and " << static_cast<int>(max) << "\n";
             } catch (...) {
                 std::cout << "Please enter a valid number.\n";
             }
